refactor: Moving MindeeSettings
