+++ src/main/java/com/mindee/MindeeSettings.java
@@ -1,4 +1,4 @@
-package com.mindee.parsing;
+package com.mindee;
 
 import lombok.Getter;
